requirements.txt
# Multi-Agent AI Research Framework
# Core dependencies for multi-agent systems

# Multi-Agent Frameworks
crewai>=0.28.0
autogen-agentchat>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# AI/ML Libraries
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Vector Databases & Embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web & API
requests>=2.31.0
httpx>=0.26.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Logging & Monitoring
loguru>=0.7.2

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
ruff>=0.1.0
